+++ spec/lib/karafka/connection/batch_delegator_spec.rb
@@ -4,43 +4,9 @@ RSpec.describe Karafka::Connection::BatchDelegator do
   subject(:delegator) { described_class }
 
   let(:group_id) { consumer_group.id }
-  let(:topic_id) { consumer_group.topics[0].name }
-  let(:topic) { Karafka::Routing::Topic.new(rand, consumer_group) }
+  let(:topic) { build(:routing_topic) }
   let(:consumer_instance) { consumer_group.topics[0].consumer.new(topic) }
-  let(:kafka_messages) { [raw_message1, raw_message2] }
-  let(:raw_message_value) { rand }
-  let(:kafka_batch) do
-    instance_double(
-      Kafka::FetchedBatch,
-      messages: kafka_messages,
-      topic: topic_id,
-      partition: 0
-    )
-  end
-  let(:raw_message1) do
-    Kafka::FetchedMessage.new(
-      message: OpenStruct.new(
-        value: raw_message_value,
-        key: nil,
-        offset: 0,
-        create_time: Time.now
-      ),
-      topic: topic_id,
-      partition: 0
-    )
-  end
-  let(:raw_message2) do
-    Kafka::FetchedMessage.new(
-      message: OpenStruct.new(
-        value: raw_message_value,
-        key: nil,
-        offset: 0,
-        create_time: Time.now
-      ),
-      topic: topic_id,
-      partition: 0
-    )
-  end
+  let(:kafka_batch) { build(:kafka_fetched_batch) }
 
   before do
     allow(Karafka::Persistence::Topic).to receive(:fetch).and_return(consumer_group.topics[0])
@@ -48,14 +14,7 @@ RSpec.describe Karafka::Connection::BatchDelegator do
   end
 
   context 'when batch_consuming true' do
-    before do
-      expect(consumer_instance)
-        .to receive(:params_batch=)
-        .with([raw_message1, raw_message2])
-
-      expect(consumer_instance)
-        .to receive(:call)
-    end
+    before { allow(consumer_instance).to receive(:call) }
 
     let(:consumer_group) do
       Karafka::Routing::Builder.instance.draw do
@@ -74,11 +33,7 @@ RSpec.describe Karafka::Connection::BatchDelegator do
   end
 
   context 'when batch_consuming false' do
-    before do
-      allow(consumer_instance).to receive(:params_batch=).with([raw_message1])
-      allow(consumer_instance).to receive(:params_batch=).with([raw_message2])
-      allow(consumer_instance).to receive(:call).twice
-    end
+    before { allow(consumer_instance).to receive(:call).twice }
 
     let(:consumer_group) do
       Karafka::Routing::Builder.instance.draw do