+++ spec/lib/karafka/connection/message_delegator_spec.rb
@@ -4,22 +4,9 @@ RSpec.describe Karafka::Connection::MessageDelegator do
   subject(:delegator) { described_class }
 
   let(:group_id) { consumer_group.id }
-  let(:topic_id) { consumer_group.topics[0].name }
-  let(:topic) { Karafka::Routing::Topic.new(rand, consumer_group) }
+  let(:topic) { build(:routing_topic) }
   let(:consumer_instance) { consumer_group.topics[0].consumer.new(topic) }
-  let(:raw_message_value) { rand }
-  let(:raw_message) do
-    Kafka::FetchedMessage.new(
-      message: OpenStruct.new(
-        value: raw_message_value,
-        key: nil,
-        offset: 0,
-        create_time: Time.now
-      ),
-      topic: topic_id,
-      partition: 0
-    )
-  end
+  let(:raw_message) { build(:kafka_fetched_message) }
   let(:consumer_group) do
     Karafka::Routing::Builder.instance.draw do
       topic :topic_name1 do
@@ -33,7 +20,6 @@ RSpec.describe Karafka::Connection::MessageDelegator do
   before do
     allow(Karafka::Persistence::Topic).to receive(:fetch).and_return(consumer_group.topics[0])
     allow(Karafka::Persistence::Consumer).to receive(:fetch).and_return(consumer_instance)
-    allow(consumer_instance).to receive(:params_batch=).with([raw_message])
     allow(consumer_instance).to receive(:call)
   end